feat(adapters): Rename ChatModel for Transformers adapter (#1012)
