+++ python/.env.example
@@ -85,7 +85,7 @@ BEEAI_LOG_LEVEL=INFO
 
 
 ########################
-### Local LLM
+### Transformers
 ########################
 
 # HF_TOKEN="your-huggingface-api-key"
