+++ python/pyproject.toml
@@ -47,6 +47,10 @@ wikipedia-api = {version = "^0.8.1", optional = true}
 beeai-sdk = {version = "^0.3.3", optional = true}
 langgraph-prebuilt = {version = "^0.6.4", optional = true}
 deprecated = "^1.2.18"
+transformers = {version = "^4.56.1", optional = true}
+torch = {version = "^2.8.0", optional = true}
+peft = {version = "^0.17.1", optional = true}
+outlines = {version = "^1.2.4", python = ">=3.11,<3.13", optional = true}
 
 [tool.poetry.group.dev.dependencies]
 commitizen = "^4.2.1"
@@ -71,7 +75,7 @@ syrupy = "^4.9.1"
 types-deprecated = "^1.2.15.20250304"
 
 [tool.poetry.extras]
-local = ["transformers", "torch", "peft", "outlines"]
+transformers = ["transformers", "torch", "peft", "outlines"]
 huggingface = ["datasets", "transformers"]
 search = ["ddgs", "wikipedia-api"]
 duckduckgo = ["ddgs"]
